[updater] feat!: upgrade VersionInfo to VersionInfoV2
In version 1 of version info, we store latest version and file list in separate
JSON files. This design retains the history version of update list, which could
use for fine update sequence. But in our ExtendedUpdater, those info is not
used.

So in version 2, we combine these two info into one JSON file to reduce network
request number. We could detect files is changed or not in just one request.

Version 1 is marked as obsolete, but it still working and could use for
compatibility.